requirements.txt
pyepics==3.4.1
PyQt5==5.14.2
PyQt5-sip==12.7.2
QtPy==1.9.0
requests==2.23.0
